change package from com.jokotech to me.thuongle
