+++ exandablerecycleradapter/src/main/AndroidManifest.xml
@@ -1,10 +1,9 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.jokotech.exandablerecycleradapter">
+    package="me.thuongle.exandablerecycleradapter">
 
     <application android:allowBackup="true"
         android:label="@string/app_name"
-        android:supportsRtl="true"
->
+        android:supportsRtl="true">
 
     </application>
