+++ recycleradapter/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="com.jokotech.recycleradapter">
+<manifest package="me.thuongle.recycleradapter">
 
     <application />
