--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/ChildItemsClickBinder.java
@@ -1,10 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-import java.util.List;
-
-public interface ChildItemsClickBinder<T> {
-
-    int getBindingVariable();
-
-    List<ClickHandler<T>> clickHandlers();
-}
