--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/ClickHandler.java
@@ -1,5 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-public interface ClickHandler<T> {
-    void onClick(T model);
-}
