--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/ItemBinder.java
@@ -1,8 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-public interface ItemBinder {
-
-    int getBindingVariable();
-
-    int getLayoutRes();
-}
