--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/ItemTouchHandler.java
@@ -1,8 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-public interface ItemTouchHandler<T> {
-
-    void onItemMove(int position, T model);
-
-    void onItemDismiss(int position, T model);
-}
