--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/LongClickHandler.java
@@ -1,5 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-public interface LongClickHandler<T> {
-    void onLongClick(T model);
-}
