--- recycleradapter/src/main/java/com/jokotech/recycleradapter/handler/OnItemTouchListener.java
@@ -1,10 +1,0 @@
-package com.jokotech.recycleradapter.handler;
-
-import android.view.View;
-
-public interface OnItemTouchListener {
-
-    boolean onItemMove(View v, int fromPosition, int toPosition);
-
-    boolean onItemDismiss(View v, int position);
-}
