+++ recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/ChildItemsClickBinder.java
@@ -1,0 +1,10 @@
+package me.thuongle.recycleradapter.handler;
+
+import java.util.List;
+
+public interface ChildItemsClickBinder<T> {
+
+    int getBindingVariable();
+
+    List<ClickHandler<T>> clickHandlers();
+}
