+++ recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/ClickHandler.java
@@ -1,0 +1,5 @@
+package me.thuongle.recycleradapter.handler;
+
+public interface ClickHandler<T> {
+    void onClick(T model);
+}
