+++ recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/ItemBinder.java
@@ -1,0 +1,8 @@
+package me.thuongle.recycleradapter.handler;
+
+public interface ItemBinder {
+
+    int getBindingVariable();
+
+    int getLayoutRes();
+}
