+++ recycleradapter/src/main/java/me/thuongle/recycleradapter/handler/ItemTouchHandler.java
@@ -1,0 +1,8 @@
+package me.thuongle.recycleradapter.handler;
+
+public interface ItemTouchHandler<T> {
+
+    void onItemMove(int position, T model);
+
+    void onItemDismiss(int position, T model);
+}
