+++ recycleradapter/src/main/java/me/thuongle/recycleradapter/util/Utils.java
@@ -1,0 +1,22 @@
+package me.thuongle.recycleradapter.util;
+
+import android.content.Context;
+import android.content.res.Resources;
+import android.support.annotation.AttrRes;
+import android.support.annotation.ColorInt;
+import android.support.annotation.NonNull;
+import android.util.TypedValue;
+
+public class Utils {
+
+    public static int dpToPx(int dp) {
+        return (int) (dp * Resources.getSystem().getDisplayMetrics().density);
+    }
+
+    @ColorInt
+    public static int getThemeColor(@NonNull final Context context, @AttrRes final int attributeColor) {
+        final TypedValue value = new TypedValue();
+        context.getTheme().resolveAttribute(attributeColor, value, true);
+        return value.resourceId;
+    }
+}
