--- sample/src/main/java/com/jokotech/adaptersample/data/TaskRepository.java
@@ -1,15 +1,0 @@
-package com.jokotech.adaptersample.data;
-
-import java.util.List;
-
-import rx.Observable;
-
-/**
- * Created by thuongleit on 8/11/16.
- */
-public interface TaskRepository {
-
-    Observable<List<Task>> getTasks();
-
-    Observable<Task> createTask();
-}
