+++ sample/src/main/java/me/thuongle/adaptersample/data/Task.java
@@ -1,0 +1,18 @@
+package me.thuongle.adaptersample.data;
+
+/**
+ * Created by thuongleit on 8/11/16.
+ */
+public class Task implements Cloneable {
+
+    public long id;
+
+    public String title;
+
+    public String description;
+
+    @Override
+    public Object clone() throws CloneNotSupportedException {
+        return super.clone();
+    }
+}
