+++ sample/src/main/java/me/thuongle/adaptersample/data/TaskRepository.java
@@ -1,0 +1,15 @@
+package me.thuongle.adaptersample.data;
+
+import java.util.List;
+
+import rx.Observable;
+
+/**
+ * Created by thuongleit on 8/11/16.
+ */
+public interface TaskRepository {
+
+    Observable<List<Task>> getTasks();
+
+    Observable<Task> createTask();
+}
