+++ sample/src/main/res/layout/content_main.xml
@@ -9,11 +9,11 @@
 
         <variable
             name="presenter"
-            type="com.jokotech.adaptersample.view.main.PresenterImpl" />
+            type="me.thuongle.adaptersample.view.main.PresenterImpl" />
 
         <variable
             name="viewModel"
-            type="com.jokotech.adaptersample.view.main.MainViewModel" />
+            type="me.thuongle.adaptersample.view.main.MainViewModel" />
 
     </data>
 
@@ -23,7 +23,7 @@
         android:paddingBottom="@dimen/activity_vertical_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
         app:layout_behavior="@string/appbar_scrolling_view_behavior"
-        tools:context=".view.main.MainActivity"
+        tools:context="me.thuongle.adaptersample.view.main.MainActivity"
         tools:showIn="@layout/activity_main">
 
         <android.support.v7.widget.RecyclerView
