+++ sample/src/main/res/layout/item_task.xml
@@ -6,11 +6,11 @@
 
         <import type="java.util.List" />
 
-        <import type="com.jokotech.recycleradapter.handler.ClickHandler" />
+        <import type="me.thuongle.recycleradapter.handler.ClickHandler" />
 
         <variable
             name="item"
-            type="com.jokotech.adaptersample.data.Task" />
+            type="me.thuongle.adaptersample.data.Task" />
 
         <variable
             name="childHandlers"
